requirements.txt
httpx
mcp[cli]
resend
python-dotenv
